+++ src/simple_duck_ml/loss/mse_loss.py
@@ -5,10 +5,10 @@ import numpy as np
 class MSELoss(ILoss):
     name = "mse"
 
-    def __call__(self, y_pred: NDArray[np.float64], y_true: NDArray[np.float64]) -> float:
+    def __call__(self, y_pred: NDArray[np.float32], y_true: NDArray[np.float32]) -> float:
         return float(np.mean((y_pred - y_true) ** 2))
 
-    def derivative(self, y_pred: NDArray[np.float64], y_true: NDArray[np.float64]) -> NDArray[np.float64]:
+    def derivative(self, y_pred: NDArray[np.float32], y_true: NDArray[np.float32]) -> NDArray[np.float32]:
         y_true = y_true.reshape(y_pred.shape)
         return 2 * (y_pred - y_true) / y_true.size
